refactor: move the definition of CobolFileKeyInfo
